requirement.txt
elasticsearch
requests
PyYAML
